Add files via upload
A lower triangular matrix is a square matrix in which all the elements above the diagonal are zero. That is, all the non-zero elements are in the lower triangle.
The first integer corresponds to the number of rows and the second number corresponds to the number of columns in the matrix. The remaining integers correspond to the elements in the matrix. The elements are read in row-wise order, first row first, then second row and so on.
Print "yes" if it is a lower triangular matrix. Print "no" if it is not a lower triangular matrix.